gas updates

+++ crates/types/src/orders/orderpool/updated_gas.rs
@@ -6,8 +6,8 @@ use crate::primitive::PoolId;
 pub struct UpdatedGas {
     pub block_number:      u64,
     pub pool_id:           PoolId,
-    pub gas_internal_book: u64,
-    pub gas_external_book: u64,
-    pub gas_internal_tob:  u64,
-    pub gas_external_tob:  u64
+    pub gas_internal_book: u128,
+    pub gas_external_book: u128,
+    pub gas_internal_tob:  u128,
+    pub gas_external_tob:  u128
 }
